WIP refactoring of validation scripts
