+++ modules/validate/checks/index.js
@@ -1,0 +1,17 @@
+const blacklist = require('./blacklist');
+const body = require('./body');
+const cmdr = require('./cmdr');
+const duplicate = require('./duplicate');
+const system = require('./system');
+const type = require('./type');
+const update = require('./update');
+
+module.exports = {
+  blacklist,
+  body,
+  cmdr,
+  duplicate,
+  system,
+  type,
+  update,
+};
